requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-api-python-client>=2.150.0
google-genai>=1.0.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
